Add two optional fields to control run-time:
- `schedule_start` (if null, defaults to "now")
- `max_run_time` (if null, defaults to `cycle_time`)

`schedule_start` allows fixing a schedule to a calendar. eg. weekly runs
at a fixed time every week.

`max_run_time` allows setting a different time for fuzzers to run, eg.
run with a `cycle_time` of 12h and `max_run_time` of 2h to run for 2h
twice a day.

This is a requirement for #45 (coverage support) and also part-time
fuzzing.