Add support for PoolConfigMap's.
This is a config that works like a map function, applying itself to each
config defined in `apply_to`. Any field can be overwritten, but some
fields must be consistent across the entire result:

- `cloud`
- `cores_per_task`
- `cpu`
- `cycle_time`
- `disk_size`
- `imageset`
- `metal`
- `minimum_memory_per_core`
- `platform`
- `schedule_start`

Also, `preprocess` fields are not supported (must be blank if used by
any config in `apply_to`).